(pp) Convert Pollenprognos adapter to TypeScript on shared base
